backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.2.0
django-filter==24.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.0.0
django-environ==0.11.2
whitenoise==6.6.0
